# swagger.yaml
openapi: 3.0.0
info:
  title: Ride Sharing API
  description: API for managing ride-sharing services
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/RS
    description: Local server
paths:
...
  /userrole/assignRole:
    post:
      summary: Assign a role to a user
      tags:
        - UserRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRole'
      responses:
        '200':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignRoleResponse'
        '404':
          description: User not found
        '500':
          description: Server error
  /userrole/getRole:
    get:
      summary: Get a user's role
      tags:
        - UserRole
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRoleResponse'
        '404':
          description: User role not found or User has no assigned role
        '500':
          description: Server error  
...
  /shareride/createRide:
    post:
      summary: Create a new shared ride
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRideRequest'
      responses:
        '201':
          description: New ride created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareRide'
...
  /shareride/searchRides:
    get:
      summary: Search for rides by startPoint, endPoint, and date
      tags:
        - ShareRide
      parameters:
        - in: query
          name: startPoint
          schema:
            type: string
        - in: query
          name: endPoint
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareRide'             
  /shareride/joinRide/{id}:
    post:
      summary: Join an existing ride
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully joined the ride
        '400':
          description: Already joined or no seats available
        '404':
          description: Ride not found
        '500':
          description: Server error
  /shareride/myCreatedRides:
    get:
      summary: Get rides created by the authenticated user
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of rides created by the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareRide'
  /shareride/myJoinedRides:
    get:
      summary: Get rides the authenticated user has joined
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of rides the user has joined
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareRide'
  /shareride/leaveRide/{id}:
    post:
      summary: Leave a shared ride
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique identifier of the ride to leave
      responses:
        '200':
          description: Successfully left the ride
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully left the ride.
        '400':
          description: Not a participant of the ride or ride is full
        '404':
          description: Ride not found
        '500':
          description: Server error
  shareride/deleteRide/{id}:
    delete:
      summary: Delete a shared ride
      tags:
        - ShareRide
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique identifier of the ride to delete
      responses:
        '200':
          description: Successfully deleted the ride
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ride was successfully deleted.
        '403':
          description: Unauthorized to delete this ride
        '404':
          description: Ride not found
        '500':
          description: Server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []